This script allows you to perform differential metabolite analysis to obtain a Log2FC, pval, padj and tval comparing two or multiple conditions.

This script allows you to perform differential metabolite analysis to obtain a Log2FC, pval, padj and tval comparing two or multiple conditions.

Usage

DMA(
  Input_data,
  Input_SettingsFile_Sample,
  Input_SettingsInfo = c(conditions = "Conditions", numerator = NULL, denominator = NULL),
  STAT_pval = "lmFit",
  STAT_padj = "fdr",
  Input_SettingsFile_Metab = NULL,
  OutputName = "",
  CoRe = FALSE,
  VST = FALSE,
  Save_as_Plot = "svg",
  Save_as_Results = "csv",
  Plot = TRUE,
  Folder_Name = NULL
)

Arguments

Input_data

DF with unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected.

Input_SettingsFile_Sample

DF which contains metadata information about the samples, which will be combined with your input data based on the unique sample identifiers used as rownames.

Input_SettingsInfo

Optional: Named vector including the information about the conditions column c(conditions="ColumnName_Plot_SettingsFile"). Can additionally pass information on numerator or denominator c(numerator = "ColumnName_Plot_SettingsFile", denominator = "ColumnName_Plot_SettingsFile") for specifying which comparison(s) will be done (one-vs-one, all-vs-one, all-vs-all). Using =NULL selects all the condition and performs multiple comparison all-vs-all. Log2FC are obtained by dividing the numerator by the denominator, thus positive Log2FC values mean higher expression in the numerator and are presented in the right side on the Volcano plot (For CoRe the Log2Distance). Default = c(conditions="Conditions", numerator = NULL, denumerator = NULL)

STAT_pval

Optional: String which contains an abbreviation of the selected test to calculate p.value. For one-vs-one comparisons choose t.test, wilcox.test, "chisq.test", "cor.test" or lmFit (=limma), for one-vs-all or all-vs-all comparison choose aov (=anova), welch(=welch anova), kruskal.test or lmFit (=limma) Default = "lmFit"

STAT_padj

Optional: String which contains an abbreviation of the selected p.adjusted test for p.value correction for multiple Hypothesis testing. Search: ?p.adjust for more methods:"BH", "fdr", "bonferroni", "holm", etc.Default = "fdr"

Input_SettingsFile_Metab

Optional: DF which contains the metadata information , i.e. pathway information, retention time,..., for each metabolite. Default = NULL

OutputName

String which is added to the output files of the DMA.

CoRe

Optional: TRUE or FALSE for whether a Consumption/Release input is used. Default = FALSE

VST

TRUE or FALSE for whether to use variance stabilizing transformation on the data when linear modeling is used for hypothesis testing. Default = TRUE

Save_as_Plot

Optional: Select the file type of output plots. Options are svg, png, pdf. Default = svg

Save_as_Results

Optional: File types for the analysis results are: "csv", "xlsx", "txt". Default = "csv"

Folder_Name

Optional: String which is added to the resulting folder name. (Default = NULL)

plot

Optional: TRUE or FALSE, if TRUE Volcano plot is saved as an overview of the results. Default = TRUE