Find additional potential IDs for "kegg", "pubchem", "chebi", "hmdb"

Find additional potential IDs for "kegg", "pubchem", "chebi", "hmdb"

Usage

EquivalentIDs(
  InputData,
  SettingsInfo = c(InputID = "MetaboliteID"),
  From = "hmdb",
  SaveAs_Table = "csv",
  FolderPath = NULL
)

Arguments

InputData

Dataframe with at least one column with the detected metabolite IDs (one ID per row).

SettingsInfo

Optional: Column name of metabolite IDs. Default = list(InputID="MetaboliteID")

From

ID type that is present in your data. Choose between "kegg", "pubchem", "chebi", "hmdb". Default = "hmdb"

SaveAs_Table

Optional: File types for the analysis results are: "csv", "xlsx", "txt". Default = "csv"

FolderPath

Optional: Path to the folder the results should be saved at. Default = NULL

Value

Input DF with additional column including potential additional IDs.

Examples

DetectedIDs <- MetaProViz::ToyData(Data="Cells_MetaData")%>% tibble::rownames_to_column("TrivialName")%>%tidyr::drop_na()
Res <- MetaProViz::EquivalentIDs(InputData= DetectedIDs, SettingsInfo = c(InputID="HMDB"), From = "hmdb")
#> Warning: The following IDs are duplicated and removed: HMDB0000725, HMDB0000267, HMDB0000755
#> chebi is used to find additional potential IDs for hmdb.