MCA_CoRe

This script allows you to perform metabolite clustering analysis and computes clusters of metabolites based on regulatory rules between Intracellular and culture media metabolomics (CoRe experiment).

Usage

MCA_CoRe(
  InputData_Intra,
  InputData_CoRe,
  SettingsInfo_Intra = c(ValueCol = "Log2FC", StatCol = "p.adj", StatCutoff = 0.05,
    ValueCutoff = 1),
  SettingsInfo_CoRe = c(DirectionCol = "CoRe", ValueCol = "Log2(Distance)", StatCol =
    "p.adj", StatCutoff = 0.05, ValueCutoff = 1),
  FeatureID = "Metabolite",
  SaveAs_Table = "csv",
  BackgroundMethod = "Intra&CoRe",
  FolderPath = NULL
)

Arguments

InputData_Intra

DF for your data (results from e.g. DMA) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns.

InputData_CoRe

DF for your data (results from e.g. DMA) containing metabolites in rows with corresponding Log2FC and stat (p-value, p.adjusted) value columns. Here we additionally require

SettingsInfo_Intra

Optional: Pass ColumnNames and Cutoffs for the intracellular metabolomics including the value column (e.g. Log2FC, Log2Diff, t.val, etc) and the stats column (e.g. p.adj, p.val). This must include: c(ValueCol=ColumnName_InputData_Intra,StatCol=ColumnName_InputData_Intra, StatCutoff= NumericValue, ValueCutoff=NumericValue) Default=c(ValueCol="Log2FC",StatCol="p.adj", StatCutoff= 0.05, ValueCutoff=1)

SettingsInfo_CoRe

Optional: Pass ColumnNames and Cutoffs for the consumption-release metabolomics including the direction column, the value column (e.g. Log2Diff, t.val, etc) and the stats column (e.g. p.adj, p.val). This must include: c(DirectionCol= ColumnName_InputData_CoRe,ValueCol=ColumnName_InputData_CoRe,StatCol=ColumnName_InputData_CoRe, StatCutoff= NumericValue, ValueCutoff=NumericValue)Default=c(DirectionCol="CoRe", ValueCol="Log2(Distance)",StatCol="p.adj", StatCutoff= 0.05, ValueCutoff=1)

FeatureID

Optional: Column name of Column including the Metabolite identifiers. This MUST BE THE SAME in each of your Input files. Default="Metabolite"

SaveAs_Table

Optional: File types for the analysis results are: "csv", "xlsx", "txt" default: "csv"

BackgroundMethod

Optional: Background method `Intra|CoRe, Intra&CoRe, CoRe, Intra or * Default="Intra&CoRe"

FolderPath

Optional: Path to the folder the results should be saved at. default: NULL

Value

MCA an instance of the MetaProViz package