Applies 80%-filtering rule, total-ion count normalization, missing value imputation and HotellingT2 outlier detection

Applies 80%-filtering rule, total-ion count normalization, missing value imputation and HotellingT2 outlier detection

Usage

PreProcessing(
  InputData,
  SettingsFile_Sample,
  SettingsInfo,
  FeatureFilt = "Modified",
  FeatureFilt_Value = 0.8,
  TIC = TRUE,
  MVI = TRUE,
  MVI_Percentage = 50,
  HotellinsConfidence = 0.99,
  CoRe = FALSE,
  SaveAs_Plot = "svg",
  SaveAs_Table = "csv",
  PrintPlot = TRUE,
  FolderPath = NULL
)

Arguments

InputData

DF which contains unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected.

SettingsFile_Sample

DF which contains information about the samples, which will be combined with the input data based on the unique sample identifiers used as rownames.

SettingsInfo

NULL or Named vector containing the information about the names of the experimental parameters. c(Conditions="ColumnName_Plot_SettingsFile", Biological_Replicates="ColumnName_Plot_SettingsFile"). Column "Conditions" with information about the sample conditions (e.g. "N" and "T" or "Normal" and "Tumor"), can be used for feature filtering and colour coding in the PCA. Column "BiologicalReplicates" including numerical values. For CoRe = TRUE a CoRe_norm_factor = "Columnname_Input_SettingsFile" and CoRe_media = "Columnname_Input_SettingsFile", have to also be added. Column CoRe_norm_factor is used for normalization and CoRe_media is used to specify the name of the media controls in the Conditions.

FeatureFilt

Optional: If NULL, no feature filtering is performed. If set to "Standard" then it applies the 80%-filtering rule (Bijlsma S. et al., 2006) on the metabolite features on the whole dataset. If is set to "Modified",filtering is done based on the different conditions, thus a column named "Conditions" must be provided in the Input_SettingsFile input file including the individual conditions you want to apply the filtering to (Yang, J et al., 2015). Default = Modified

FeatureFilt_Value

Optional: Percentage of feature filtering. Default = 0.8

TIC

Optional: If TRUE, Total Ion Count normalization is performed. Default = TRUE

MVI

Optional: If TRUE, Missing Value Imputation (MVI) based on half minimum is performed Default = TRUE

MVI_Percentage

Optional: Percentage 0-100 of imputed value based on the minimum value. Default = 50

HotellinsConfidence

Optional: Defines the Confidence of Outlier identification in HotellingT2 test. Must be numeric.Default = 0.99

CoRe

Optional: If TRUE, a consumption-release experiment has been performed and the CoRe value will be calculated. Please consider providing a Normalisation factor column called "CoRe_norm_factor" in your "Input_SettingsFile" DF, where the column "Conditions" matches. The normalisation factor must be a numerical value obtained from growth rate that has been obtained from a growth curve or growth factor that was obtained by the ratio of cell count/protein quantification at the start point to cell count/protein quantification at the end point.. Additionally control media samples have to be available in the "Input" DF and defined as "CoRe_media" samples in the "Conditions" column in the "Input_SettingsFile" DF. Default = FALSE

SaveAs_Plot

Optional: Select the file type of output plots. Options are svg, png, pdf. If set to NULL, plots are not saved. Default = svg

SaveAs_Table

Optional: Select the file type of output table. Options are "csv", "xlsx", "txt". If set to NULL, plots are not saved. Default = "csv"

PrintPlot

Optional: If TRUE prints an overview of resulting plots. Default = TRUE

FolderPath

Optional: Path to the folder the results should be saved at. default: NULL