Merges the analytical replicates of an experiment

Merges the analytical replicates of an experiment

Usage

ReplicateSum(
  InputData,
  SettingsFile_Sample,
  SettingsInfo = c(Conditions = "Conditions", Biological_Replicates =
    "Biological_Replicates", Analytical_Replicates = "Analytical_Replicates"),
  SaveAs_Table = "csv",
  FolderPath = NULL
)

Arguments

InputData

DF which contains unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected. #@param SettingsFile_Sample DF which contains information about the samples Column "Conditions", "Biological_replicates" and "Analytical_Replicates has to exist.

SettingsInfo

Optional: Named vector including the Conditions and Replicates information: c(Conditions="ColumnNameConditions", Biological_Replicates="ColumnName_SettingsFile_Sample", Analytical_Replicates="ColumnName_SettingsFile_Sample").Default = NULL

SaveAs_Table

Optional: File types for the analysis results are: "csv", "xlsx", "txt", ot NULL default: "csv"

FolderPath

Optional: Path to the folder the results should be saved at. default: NULL