StandardORA

Uses enricher to run ORA on the differential metabolites (DM) using a pathway list

Usage

StandardORA(
  InputData,
  SettingsInfo = c(pvalColumn = "p.adj", PercentageColumn = "t.val", PathwayTerm =
    "term", PathwayFeature = "Metabolite"),
  pCutoff = 0.05,
  PercentageCutoff = 10,
  PathwayFile,
  PathwayName = "",
  minGSSize = 10,
  maxGSSize = 1000,
  SaveAs_Table = "csv",
  FolderPath = NULL
)

Arguments

InputData

DF with metabolite names/metabolite IDs as row names. Metabolite names/IDs need to match the identifier type (e.g. HMDB IDs) in the PathwayFile.

SettingsInfo

Optional: Pass ColumnName of the column including parameters to use for pCutoff and PercentageCutoff. Also pass ColumnName for PathwayFile including term and feature names. (pvalColumn = ColumnName InputData, PercentageColumn= ColumnName InputData, PathwayTerm= ColumnName PathwayFile, PathwayFeature= ColumnName PathwayFile) c(pvalColumn="p.adj", PercentageColumn="t.val", PathwayTerm= "term", PathwayFeature= "Metabolite")

pCutoff

Optional: p-adjusted value cutoff from ORA results. Must be a numeric value. default: 0.05

PercentageCutoff

Optional: Percentage cutoff of metabolites that should be considered for ORA. Selects Top/Bottom % of selected PercentageColumn, usually t.val or Log2FC default: 10

PathwayFile

DF that must include column "term" with the pathway name, column "Metabolite" with the Metabolite name or ID and column "Description" with pathway description that will be depicted on the plots.

PathwayName

Optional: Name of the PathwayFile used default: ""

minGSSize

Optional: minimum group size in ORA default: 10

maxGSSize

Optional: maximum group size in ORA default: 1000

FolderPath

Optional: Path to the folder the results should be saved at. default: NULL

Save_as_Results

Optional: File types for the analysis results are: "csv", "xlsx", "txt" default: "csv"

Value

Saves results as individual .csv files.