This script allows you to perform different data visualizations using the results of the MetaProViz analysis

This script allows you to perform different data visualizations using the results of the MetaProViz analysis

Usage

VizVolcano(
  Plot_Settings = "Standard",
  Plot_SettingsInfo = NULL,
  Plot_SettingsFile = NULL,
  Input_data,
  y = "p.adj",
  x = "Log2FC",
  AdditionalInput_data = NULL,
  OutputPlotName = "",
  Comparison_name = c(Input_data = "Cond1", AdditionalInput_data = "Cond2"),
  xlab = NULL,
  ylab = NULL,
  pCutoff = 0.05,
  FCcutoff = 0.5,
  color_palette = NULL,
  shape_palette = NULL,
  SelectLab = "",
  Connectors = FALSE,
  Subtitle = "",
  Theme = NULL,
  Save_as_Plot = "svg",
  Folder_Name = NULL
)

Arguments

Plot_Settings

Optional: Choose between "Standard" (Input_data), "Compare" (plot two comparisons together Input_data and Input_data2) or "PEA" (Pathway Enrichment Analysis) Default = "Standard"

Plot_SettingsInfo

Optional: NULL or Named vector including at least one of those three information for Plot_Settings="Standard" or "Compare": c(color="ColumnName_Plot_SettingsFile", shape= "ColumnName_Plot_SettingsFile", individual="ColumnName_Plot_SettingsFile"). For Plot_Settings="PEA" a named vector with c(PEA_Pathway="ColumnNameAdditionalInput_data", PEA_score="ColumnNameAdditionalInput_data", PEA_stat= "ColumnNameAdditionalInput_data", individual="Plot_SettingsFile), optionally you can additionally include c(color="ColumnName_Plot_SettingsFile", shape= "ColumnName_Plot_SettingsFile").Default = NULL

Plot_SettingsFile

Optional: DF with column "Metabolite" including the Metabolite names (needs to match Metabolite names of Input_data) and other columns with required PlotSettingInfo. Default = NULL

Input_data

DF with column "Metabolite" including the Metabolite names, Log2FC, pvalue/padjusted values. Can also include additional columns with metadata usable for Plot_Setting_Info.

y

Optional: Column name including the values that should be used for y-axis. Usually this would include the p.adjusted value. Default = "p.adj"

x

Optional: Column name including the values that should be used for x-axis. Usually this would include the Log2FC value. Default = "Log2FC"

AdditionalInput_data

Optional: DF to compare to main Input_data with the same column names x and y (Plot_Settings="Compare") or Pathway enrichment analysis results (Plot_Settings="PEA"). Default = NULL

OutputPlotName

Optional: String which is added to the output files of the plot. Default = ""

Comparison_name

Optional: Named vector including those information about the two datasets that are compared on the plots when choosing Plot_Settings= "Compare". Default = c(Input_data="Cond1", AdditionalInput_data= "Cond2")

xlab

Optional: String to replace x-axis label in plot. Default = NULL

ylab

Optional: String to replace y-axis label in plot. Default = NULL

pCutoff

Optional: Number of the desired p value cutoff for assessing significance. Default = 0.05

FCcutoff

Optional: Number of the desired log fold change cutoff for assessing significance. Default = 0.5

color_palette

Optional: Provide customiced color-palette in vector format. Default = NULL

shape_palette

Optional: Provide customiced shape-palette in vector format. Default = NULL

SelectLab

Optional: If set to NULL, feature labels will be plotted randomly. If vector is provided, e.g. c("MetaboliteName1", "MetaboliteName2"), selected names will be plotted. If set to default "", no feature names will be plotted. Default = ""

Connectors

Optional: TRUE or FALSE for whether Connectors from names to points are to be added to the plot. Default = FALSE

Subtitle

Optional: Default = ""

Theme

Optional: Selection of theme for plot, e.g. theme_grey(). You can check for complete themes here: https://ggplot2.tidyverse.org/reference/ggtheme.html. Default = NULL

Save_as_Plot

Optional: Select the file type of output plots. Options are svg, pdf, png or NULL. Default = "svg"

Folder_Name

Optional: String which is added to the resulting folder name (Default = NULL)