Check and summarize prior_knowledge-to-MeasuredFeatures relationship

Check and summarize prior_knowledge-to-MeasuredFeatures relationship

Usage

checkmatch_pk_to_data(
  data,
  input_pk,
  metadata_info = c(InputID = "HMDB", PriorID = "HMDB", grouping_variable = "term"),
  save_table = "csv",
  path = NULL
)

Arguments

data

dataframe with at least one column with the detected metabolite IDs (e.g. HMDB). If there are multiple IDs per detected peak, please separate them by comma ("," or ", " or chr list). If there is a main ID and additional IDs, please provide them in separate columns.

input_pk

dataframe with at least one column with the metabolite ID (e.g. HMDB) that need to match data metabolite IDs "source" (e.g. term). If there are multiple IDs, as the original pathway IDs (e.g. KEGG) where translated (e.g. to HMDB), please separate them by comma ("," or ", " or chr list).

metadata_info

Colum name of Metabolite IDs in data and input_pk as well as column name of grouping_variable in input_pk. Default = c(InputID="HMDB", PriorID="HMDB", grouping_variable="term")

save_table

Optional: File types for the analysis results are: "csv", "xlsx", "txt". Default = "csv"

path

Optional: Path to the folder the results should be saved at. Default = NULL

Examples

DetectedIDs <-  cellular_meta %>%dplyr::select("Metabolite", "HMDB")%>%tidyr::drop_na()
input_pathway <- MetaProViz::translate_id(data= MetaProViz::metsigdb_kegg(), metadata_info = c(InputID="MetaboliteID", grouping_variable="term"), from = c("kegg"), to = c("hmdb"))[["TranslatedDF"]]%>%tidyr::drop_na()
Res <- MetaProViz::checkmatch_pk_to_data(data= DetectedIDs, input_pk= input_pathway, metadata_info = c(InputID="HMDB", PriorID="hmdb", grouping_variable="term"))
#> Warning: 4 duplicated IDs were removed from column HMDB
#> Warning: 8802 duplicated IDs were removed from PK column hmdb
#> Error in mutate(., Count_FeatureIDs_to_GroupingVariable = case_when(is.na(!!sym(metadata_info[["grouping_variable"]])) ~     NA_integer_, TRUE ~ n_distinct(!!sym(metadata_info[["PriorID"]]),     na.rm = TRUE))): ℹ In argument: `Count_FeatureIDs_to_GroupingVariable = case_when(...)`.
#> ℹ In group 1: `HMDB = "HMDB0000001"` `term = "Histidine metabolism, Metabolic
#>   pathways"`.
#> Caused by error in `case_when()`:
#> ! Failed to evaluate the right-hand side of formula 2.
#> Caused by error in `n_distinct()`:
#> ! could not find function "n_distinct"