Translate IDs

Translate IDs to and from KEGG, PubChem, Chebi.

Usage

TranslateID(
  Input_DataFrame,
  SettingsInfo = list(IdColumn = "MetaboliteID", FromFormat = c("kegg"), ToFormat =
    c("pubchem", "chebi", "hmdb"), Method = "GetAll", GroupingVariable = "term")
)

Arguments

Input_DataFrame

Dataframe with two columns for source (=term) and Target (=gene), e.g. Hallmarks.

SettingsInfo

Optional: Column name of Target in Input_GeneSet. Default = list(IdColumn="MetaboliteID", FromFormat=c("kegg"), ToFormat=c("pubchem","chebi","hmdb"), Method="GetAll", GroupingVariable="term")

SaveAs_Table

Optional: File types for the analysis results are: "csv", "xlsx", "txt". Default = "csv"

FolderPath

Optional: String which is added to the resulting folder name Default = NULL

Value

3 data frames: 1) Original data and the new column of translated ids. 2) Mapping summary from Original ID to Translated. 3) Mapping summary from Translated to Original.