This script allows you to perform PCA plot visualization using the results of the MetaProViz analysis

This script allows you to perform PCA plot visualization using the results of the MetaProViz analysis

Usage

VizPCA(
  Plot_SettingsInfo = NULL,
  Plot_SettingsFile = NULL,
  Input_data,
  color_palette = NULL,
  color_scale = "discrete",
  shape_palette = NULL,
  Show_Loadings = FALSE,
  Scaling = TRUE,
  Theme = NULL,
  OutputPlotName = "",
  Save_as_Plot = "svg",
  Folder_Name = NULL
)

Arguments

Plot_SettingsInfo

Optional: NULL or Named vector including at least one of those three information : c(color="ColumnName_Plot_SettingsFile", shape= "ColumnName_Plot_SettingsFile"). Default = NULL

Plot_SettingsFile

Optional: DF which contains information about the samples, which will be combined with your input data based on the unique sample identifiers used as rownames. Column "Conditions" with information about the sample conditions (e.g. "N" and "T" or "Normal" and "Tumor"), can be used for feature filtering and colour coding in the PCA. Column "AnalyticalReplicate" including numerical values, defines technical repetitions of measurements, which will be summarised. Column "BiologicalReplicates" including numerical values. Please use the following names: "Conditions", "Biological_Replicates", "Analytical_Replicates".Default = NULL

Input_data

DF with unique sample identifiers as row names and metabolite numerical values in columns with metabolite identifiers as column names. Use NA for metabolites that were not detected. includes experimental design and outlier column.

color_palette

Optional: Provide customiced color-palette in vector format. For continuous scale use e.g. scale_color_gradient(low = "#88CCEE", high = "red") and for discrete scale c("#88CCEE", "#DDCC77","#661100", "#332288")Default = NULL

color_scale

Optional: Either "continuous" or "discrete" colour scale. For numeric or integer you can choose either, for character you have to choose discrete. Default = NULL

shape_palette

Optional: Provide customiced shape-palette in vector format. Default = NULL

Show_Loadings

Optional: TRUE or FALSE for whether PCA loadings are also plotted on the PCA (biplot) Default = FALSE

Scaling

Optional: TRUE or FALSE for whether a data scaling is used Default = TRUE

Theme

Optional: Selection of theme for plot, e.g. theme_grey(). You can check for complete themes here: https://ggplot2.tidyverse.org/reference/ggtheme.html. If default=NULL we use theme_classic(). Default = "discrete"

OutputPlotName

Optional: String which is added to the output files of the PCA Default = ""

Save_as_Plot

Optional: Select the file type of output plots. Options are svg, png, pdf or NULL. Default = svg

Folder_Name

Optional: String which is added to the resulting folder name (Default = NULL)